Fixes for cuda to work with Ubuntu 16 cuda package. Changed flycap interface to eliminate namespace conflicts with libdc1394 interface.

+++ include/typedefs.h
@@ -13,7 +13,8 @@
 #if HAVE_SYS_TYPES_H
 #include <sys/types.h>
 #else
-typedef uint32_t u_long;
+//typedef uint32_t u_long;
+typedef unsigned long u_long;
 typedef unsigned int u_int;
 #endif
